Removed unused variables. Switch hard-coded parameters to more (!) routine arguments for all files. Small unexplained feature with fitrad in allstar removed because not present in other daophot sources codes elsewhere (i suspect we got a debugging version). Modified Files:  	addstr.f allstr.f daotools.f find.f fotometry.f fudge.f  	getpsf.f group.f iosubs.f mathsubs.f nstar.f pckpsf.f peak.f  	sort.f substar.f

+++ dao/fotometry.f
@@ -462,7 +462,7 @@ C (3) the uncertainty of the mean sky brightness (this standard error
 C increases directly with the area of the aperture).
 C
       MAGERR(I)=1.0857*SQRT(ERROR(1)+ERROR(2)+ERROR(3))/SNGL(APMAG(I))
-      IF (I .GT. 1) MAGERR(I) = MAGERR(I) * SQRT(AREA(I)/BTOT(I))
+      IF (I .GT. 1) MAGERR(I) = MAGERR(I) * SQRT(REAL(AREA(I))/BTOT(I))
       MAGERR(I) = AMIN1(9.9999, MAGERR(I))
       APMAG(I)=25.D0-2.5D0*DLOG10(APMAG(I))
       IF (APMAG(I) .GT. 94.999D0) THEN
